backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
SQLAlchemy==2.0.44
psycopg[binary]==3.2.12
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
pydantic==2.11.9
pydantic-settings==2.11.0
httpx==0.28.1
slowapi==0.1.9
redis==6.4.0
pytest==8.4.2

